Allow for strings in 'only' and 'except' options
With the addition of the `only` and `except` options, it opens up the
possiblities to the [partial response](http://bit.ly/1fxWuXJ) design
pattern where the consumer of the API can specifically request only the
fields that it desires. A common pattern to request these fields would
be via `params`, which will arrive as strings. Currently only an array
of symbols (or hash containing symbols) is supported to define the
desired fields.

This commit updates Grape::Entity to support an array of symbols _or_
strings to be passed as the `only` or `except` options of `represent`.