Rewrite array methods delegation to @exposures.
ActiveSupport's `delegate` adds an overhead that we simply don't need.
And it's easy to implement efficient delegation for this specific task.
This patch also adds new methods for delegation: `select`, `all?` and
`each_with_object` that are heavily used internally.

It's a micro-optimization but it's easy to implement so here it is.