support dynamic attributes & merging collection values
Work in progress. Allows the parent attribute of a nested attribute
structure to be a dynamic attribute that is evaluated during
rendering. Adds a 'combine' option that can be used when exposing a
collection of items in a single presenter that controls how the
individual items are combined. No value maintains the existing behavior,
where each item is rendered as JSON and appended to the Array, while a
value of :merge results in each item's JSON representation being
Hash#merge'd into a single Hash.