Support hashes with string keys
Before this change the only way I could find to access a hash with string keys
was to either convert it to a symbol hash or define methods in the entity.

This should at least allow the entire hash to be accessed with string keys yet default to existing behaviour (symbol keys).